Fix calculate cost function to return float

+++ lab04/agm.hpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-int calculate_cost(vector<NodeSource> adjacency);
+float calculate_cost(vector<NodeSource> adjacency);
 vector< pair< pair<int, int>, int > > create_adjacency_edges(vector<NodeSource> adjacency);
 vector<NodeSource> transform(vector<NodeSource> adjacency, vector< pair< pair<int, int>, int > > agm_edges, int nodes);
 bool sort_cost_ascending(const pair< pair<int, int>, int > &rhs, const pair< pair<int, int>, int > &lhs);
